requirements.txt
requests
selenium
pykakasi
python-dotenv
beautifulsoup4
